:recycle: split Unicode tests
